One symbol per file, folders by kind, barrels throughout
Two files were carrying everything. Now constants live at the root of src,
the handlers hold nothing but the lambda entry point, and types, schemas,
services and helpers each get a folder with a barrel and one symbol per
file. The tests mirror it.

handlers is the one folder without a barrel, because esbuild takes every
file in there as a lambda entry point and a barrel would land a third
artifact in the zip. Nothing imports handlers anyway, AWS reaches them by
file name.

The split earned its keep straight away. isNotFoundInZuora and
undefinedIfNotFound had only ever been exercised through
processPaymentMethod and now have direct tests, which is where the
narrowness of the error-code check belongs. And the scrubBatch tests had
been quietly inheriting stubs from a shared beforeEach; on their own they
have to declare what they depend on, which is the point.

Seventeen unit tests where there were twelve. The integration test still
scrubs a real card in CODE, and the build still produces exactly the two
artifacts the CDK names.
