+++ handlers/scrub-non-tokenised-payment-methods/src/schemas/index.ts
@@ -1,0 +1,2 @@
+export { nonTokenisedCardsSchema } from './nonTokenisedCardsSchema';
+export { subscriptionStatusSchema } from './subscriptionStatusSchema';
