+++ handlers/scrub-non-tokenised-payment-methods/src/services/index.ts
@@ -1,0 +1,3 @@
+export { findPaymentMethodsToScrub } from './findPaymentMethodsToScrub';
+export { processPaymentMethod } from './processPaymentMethod';
+export { scrubBatch } from './scrubBatch';
