+++ handlers/scrub-non-tokenised-payment-methods/src/types/batchResult.ts
@@ -1,0 +1,6 @@
+/** The tally a batch reports back to the state machine. */
+export type BatchResult = {
+	scrubbed: number;
+	wouldScrub: number;
+	skipped: number;
+};
