+++ handlers/scrub-non-tokenised-payment-methods/src/types/index.ts
@@ -1,0 +1,4 @@
+export type { BatchResult } from './batchResult';
+export type { LambdaEvent } from './lambdaEvent';
+export type { Outcome } from './outcome';
+export type { PaymentMethodToScrub } from './paymentMethodToScrub';
