+++ handlers/scrub-non-tokenised-payment-methods/src/types/lambdaEvent.ts
@@ -1,0 +1,6 @@
+import type { PaymentMethodToScrub } from './paymentMethodToScrub';
+
+/** What the distributed map hands to the scrubbing lambda. */
+export type LambdaEvent = {
+	Items: PaymentMethodToScrub[];
+};
