+++ handlers/scrub-non-tokenised-payment-methods/src/types/outcome.ts
@@ -1,0 +1,2 @@
+/** What happened to a single payment method. */
+export type Outcome = 'scrubbed' | 'skipped' | 'wouldScrub';
