+++ handlers/scrub-non-tokenised-payment-methods/test/fixtures/creditCardFixture.ts
@@ -1,0 +1,5 @@
+export const creditCardFixture = (overrides: Record<string, unknown> = {}) => ({
+	id: 'pm-1',
+	status: 'Active',
+	...overrides,
+});
