+++ handlers/scrub-non-tokenised-payment-methods/test/fixtures/paymentMethodToScrubFixture.ts
@@ -1,0 +1,7 @@
+import type { PaymentMethodToScrub } from '../../src/types';
+
+export const paymentMethodToScrubFixture: PaymentMethodToScrub = {
+	payment_method_id: 'pm-1',
+	account_id: 'acc-id-1',
+	account_number: 'A-S00000001',
+};
