+++ handlers/scrub-non-tokenised-payment-methods/test/fixtures/subscriptionFixtures.ts
@@ -1,0 +1,2 @@
+export const cancelledSubscription = { status: 'Cancelled' };
+export const activeSubscription = { status: 'Active' };
