+++ handlers/scrub-non-tokenised-payment-methods/test/fixtures/zuoraErrorFixture.ts
@@ -1,0 +1,8 @@
+import { ZuoraError } from '@modules/zuora/errors';
+
+export const zuoraErrorFixture = (code: string, message: string): ZuoraError =>
+	new ZuoraError(
+		message,
+		{ status: 200, responseBody: '', responseHeaders: {} },
+		[{ code, message }],
+	);
